[Strategy Types] Add some logic to StrategyType enum
This allows us to weigh algorithms differently to be able to combine
them all together. I don't know how to weigh them; that is part of the
tuning in real life tests.

The HYBRID algorithm should always be weighed 0 in order to prevent it
from instantiating and calling itself over and over again.